add StorageLocation properties
During RDSS-Archivematica Alpha MVP Sprint 2, we encountered the requirement to retrieve all the Files that are included for a given Dataset from their JiscRDSS cloud storage location (S3). These Files are listed in the Metadata Read message payload that RDSS-Archivematica receives to trigger a register-preservation-message event for that Dataset. However, the current version of the RDSS Canonical Data Model (CDM) does not have the necessary properties specified to use for file storage location values.