[RDSS-1861] Push generated documentation to master if running from non auto-generated master push
If _not_ on master, then the `git push --dry-run` would still check if the push
would fail. This is good, because it would catch cases where someone else has
pushed something in the meantime.

We are deliberatly not using the Travis [s-k-i-p c-i] method: this is for PRs
master -> develop, which is how version bumps are done. The last commit is an
auto generated one, but we still want Travis so Travis can approve the PR.

(The hyphens above are so Travis does _not_ skip the build for _this_ commit!)