[RDSS-1861] Run pre-commit on generated files to fix end of files
This is since by default the generated files make pre-commit fail, which means,
if the files are committed without pre-commit changing the files, on the _next_
PR, the build will fail because pre-commit is run before anything else.

Suspect not doing this in the earlier version of the automatic generation
script was a contributing factor to merge conflicts.